requirements.txt
# ============================================================
# telecom-intelligence · Python dependencies
# ============================================================

# Phase 1: Synthetic Data Generation
networkx>=3.2
pyyaml>=6.0

# Testing
pytest>=8.0
